requirements.txt
autopep8==1.7.0
numpy==1.23.3
opencv-python==4.6.0.66
pycodestyle==2.9.1
python-dotenv==0.21.0
pyzmq==24.0.0
toml==0.10.2
